Update GitHub repository URL
Because you renamed the project, the repository URL for the project has
changed. In the browser, GitHub properly redirects to the new repo, but
if a user has the GitHub app installed, it simply says "Loading
repository failed!" because it simply takes the URL and parses that into
an API request. The GitHub API does not handle repository rename
redirects.

See:
http://stackoverflow.com/questions/22715169/github-api-returns-404-but-default-http-request-returns-301
https://github.com/github/android/blob/e18693727bc62bc2598cd7575130cbf3b8f341eb/app/src/main/java/com/github/mobile/ui/repo/RepositoryViewActivity.java#L121
https://github.com/github/android/blob/e18693727bc62bc2598cd7575130cbf3b8f341eb/app/src/main/java/com/github/mobile/core/repo/RefreshRepositoryTask.java#L55
https://github.com/eclipse/egit-github/blob/bcdf39cca5ef0dd828f960030a2febba1e56a253/org.eclipse.egit.github.core/src/org/eclipse/egit/github/core/service/RepositoryService.java#L580-L607